Cleaned up the code for Time module and added unit test to module source as a sub module.
